separate out strudel.mjs, make index.mjs aggregate module
