+++ packages/midi/midi.mjs
@@ -1,6 +1,6 @@
 import { isNote } from 'tone';
 import _WebMidi from 'webmidi';
-import { Pattern, isPattern } from '@strudel.cycles/core/strudel.mjs';
+import { Pattern, isPattern } from '@strudel.cycles/core';
 import { Tone } from '@strudel.cycles/tone';
 
 // if you use WebMidi from outside of this package, make sure to import that instance:
