+++ packages/tonal/voicings.mjs
@@ -1,4 +1,4 @@
-import { Pattern as _Pattern, stack, Hap, reify } from '@strudel.cycles/core/strudel.mjs';
+import { Pattern as _Pattern, stack, Hap, reify } from '@strudel.cycles/core';
 import _voicings from 'chord-voicings';
 const { dictionaryVoicing, minTopNoteDiff, lefthand } = _voicings.default || _voicings; // parcel module resolution fuckup
